+++ application/views/template.php
@@ -17,7 +17,7 @@
 	<div id="navigation">
 		<ul>
 			<li><a href="/" {dActive}>Dashboard</a></li>
-			<li><a href="/admin" {aActive}>Administration</a></li>
+			<li><a href="/admin" {aActive}>Maintenance</a></li>
 			<li><a href="/receiving" {rActive}>Receiving</a></li>
 			<li><a href="/production" {pActive}>Production</a></li>
 			<li><a href="/sales" {sActive}>Sales</a></li>