+++ Changelog.md
@@ -3,6 +3,9 @@
 ============================================
 Members: Matthew, Dennis, Injo, Haram, Lydia
 
+2016-12-08 12:55 - Dennis
+- Maintenance controller materials read complete
+
 2016-12-08 00:42 - Dennis
 -Consolidated recipe and product tables, renamed materialcombo table to recipe
 -Replaced missing libraries and helpers in autoload